+++ webapp/javascript/ui/Portal.tsx
@@ -1,0 +1,3 @@
+import Portal, { PortalProps } from '@mui/base/Portal';
+
+export { Portal, PortalProps };